pyproject.toml
[project]
name = "ace-active-carver"
version = "0.1.0"
description = "Uncertainty-driven on-the-fly active learning system using Pacemaker (ACE potentials) and LAMMPS."
readme = "README.md"
requires-python = ">=3.10,<3.11"
dependencies = [
    "pacemaker",
    "ase",
    "numpy",
    "pandas",
    "scipy",
    "lammps",
    "PyYAML",
    "mace-torch",
    "pyxtal",
    "scikit-learn",
    "pytest>=9.0.1",
    "pymatgen",
    "mp-api",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[dependency-groups]
dev = [
    "black",
    "isort",
    "pydocstyle",
    "pytest",
]

[tool.black]
line-length = 88
target-version = ['py310']

[tool.isort]
profile = "black"
